%d en cours

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: nrochard <nrochard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 22:22:52 by nrochard          #+#    #+#             */
-/*   Updated: 2019/12/06 21:04:46 by nrochard         ###   ########.fr       */
+/*   Updated: 2019/12/16 00:02:41 by nrochard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int a;
